Create README.md
Hi I am @zabby1050'
I am a positivist and an optimism about the future of the world and the betterment of human life.
I am learning coding and computer programming to enable me to fit in the economic dynamics of the digital world and business. I am interested in cryptocurrency, digital marketing and business management as a broad basis for my future career. I am looking to collaborate with any entity, business or person and enterprises that pursue common interests. GOOD LUCK everyone lets pull each other to the front as we build each other mutually as we grow.